+++ CHANGELOG.md
@@ -10,11 +10,11 @@ this project uses date-based 'snapshot' version identifiers.
 ### Added
 
 - Switch `-m|--message` for providing upload announcement from command line
+- Switch `-F|--file` for providing upload announcement from file
 
 ### Changed
 
 - Enable `--dry-run` option for `upload` target
-- Require (possibly empty) announcement text for uploads
 
 ### Fixed